+++ src/modules/all.ts
@@ -12,6 +12,7 @@ export * from './esl-toggleable/core';
 export * from './esl-popup/core';
 export * from './esl-trigger/core';
 
+export * from './esl-panel-group/core';
 export * from './esl-panel/core';
 
 export * from './esl-tab/core';