+++ src/modules/esl-panel-group/core.ts
@@ -1,0 +1,3 @@
+export type {ESLPanelGroupShape} from './core/esl-panel-group.shape';
+
+export * from './core/esl-panel-group';