+++ src/modules/esl-tab/core/esl-tab.shape.ts
@@ -1,0 +1,8 @@
+import type {ESLTriggerShape} from '../../esl-trigger/core/esl-trigger.shape';
+
+/**
+ * Tag declaration interface of {@link ESLTab} element
+ * Used for JSX declaration
+ */
+export interface ESLTabShape extends ESLTriggerShape {
+}