+++ src/modules/esl-toggleable/core/esl-toggleable.ts
@@ -24,7 +24,7 @@ export interface ToggleableActionParams {
   /** Activate hover tracking to hide Toggleable */
   trackHover?: boolean;
   /** Element activator of the action */
-  activator?: HTMLElement;
+  activator?: HTMLElement | null;
   /** Event that initiates the action */
   event?: Event;
 
@@ -82,6 +82,8 @@ export class ESLToggleable extends ESLBaseElement {
   protected _task: DelayedTask = new DelayedTask();
   /** Marker for current hover listener state */
   protected _trackHover: boolean = false;
+  /** Delay for track hover listeners actions */
+  protected _trackHoverDelay: number | undefined;
 
   protected connectedCallback() {
     super.connectedCallback();
@@ -140,8 +142,9 @@ export class ESLToggleable extends ESLBaseElement {
     }
   }
   /** Bind hover events listeners for the Toggleable itself */
-  protected bindHoverStateTracking(track: boolean) {
+  protected bindHoverStateTracking(track: boolean, hideDelay?: number | string) {
     if (!DeviceDetector.hasHover) return;
+    this._trackHoverDelay = track && hideDelay !== undefined ? +hideDelay : undefined;
     if (this._trackHover === track) return;
     this._trackHover = track;
 
@@ -168,7 +171,7 @@ export class ESLToggleable extends ESLBaseElement {
     params = this.mergeDefaultParams(params);
     this._task.put(this.showTask.bind(this, params), defined(params.showDelay, params.delay));
     this.bindOutsideEventTracking(this.closeOnOutsideAction);
-    this.bindHoverStateTracking(!!params.trackHover);
+    this.bindHoverStateTracking(!!params.trackHover, defined(params.hideDelay, params.delay));
     return this;
   }
   /** Change the element state to inactive */
@@ -176,7 +179,7 @@ export class ESLToggleable extends ESLBaseElement {
     params = this.mergeDefaultParams(params);
     this._task.put(this.hideTask.bind(this, params), defined(params.hideDelay, params.delay));
     this.bindOutsideEventTracking(false);
-    this.bindHoverStateTracking(!!params.trackHover);
+    this.bindHoverStateTracking(!!params.trackHover, defined(params.hideDelay, params.delay));
     return this;
   }
 
@@ -192,7 +195,7 @@ export class ESLToggleable extends ESLBaseElement {
   /** Actual hide task to execute by toggleable task manger ({@link DelayedTask} out of the box) */
   protected hideTask(params: ToggleableActionParams) {
     if (!params.force && !this.open) return;
-    if (!params.silent && !this.$$fire('before:hide',{detail: {params}})) return;
+    if (!params.silent && !this.$$fire('before:hide', {detail: {params}})) return;
     this.open = false;
     this.onHide(params);
     if (!params.silent) this.$$fire('hide', {detail: {params}, cancelable: false});
@@ -234,14 +237,14 @@ export class ESLToggleable extends ESLBaseElement {
     return activators.get(this);
   }
   public set activator(el: HTMLElement | null | undefined) {
-    el ? activators.set(this, el): activators.delete(this);
+    el ? activators.set(this, el) : activators.delete(this);
   }
 
   /** Returns the element to apply a11y attributes */
-  protected get $a11yTarget(): HTMLElement | undefined {
+  protected get $a11yTarget(): HTMLElement | null {
     const target = this.getAttribute('a11y-target');
-    if (target === 'none') return;
-    return target ? this.querySelector(target) as HTMLElement : this;
+    if (target === 'none') return null;
+    return target ? this.querySelector(target) : this;
   }
 
   /** Called on show and on hide actions to update a11y state accordingly */
@@ -277,12 +280,14 @@ export class ESLToggleable extends ESLBaseElement {
 
   @bind
   protected _onMouseEnter(e: MouseEvent) {
-    const baseParams = {initiator: 'mouseenter', trackHover: true, activator: this.activator, event: e};
+    const hideDelay = this._trackHoverDelay;
+    const baseParams: ToggleableActionParams = {initiator: 'mouseenter', trackHover: true, activator: this.activator, event: e, hideDelay};
     this.show(Object.assign(baseParams, this.trackHoverParams));
   }
   @bind
   protected _onMouseLeave(e: MouseEvent) {
-    const baseParams = {initiator: 'mouseleave', trackHover: true, activator: this.activator, event: e};
+    const hideDelay = this._trackHoverDelay;
+    const baseParams: ToggleableActionParams = {initiator: 'mouseleave', trackHover: true, activator: this.activator, event: e, hideDelay};
     this.hide(Object.assign(baseParams, this.trackHoverParams));
   }
 }